Update db-schema-create-and-populate.sql
This changes fixes a problem where in some versions of postgres (9.6 x64 windows, for example), the original script appears to work fine, but later inserts fail with a duplicate error because the sequences are still set to a currentValue of 1 after the table is populated with inserts that explicitly set the id rather than using the auto sequencer. This is probably a bug in postgres, but figuring it out is an hour I'll never get back, and the proposed changes fix the problem.